Split legend hooks and refresh metadata
Moves legend creation and legend interaction logic into a dedicated hook module, then re-exports it from the shared hooks entry to preserve existing imports. Also refreshes the generated API surface docs, bundle-size benchmarks, and the physics kernel comment for the gravity support check.
